Interpolate data onto surface when changing the window of interest.
Added a new function openep.case.case_routines.get_sample_indices_within_woi that is used in openep.case.case_routines.calculate_voltage_from_surface.

This correctly accounts for the fact that the window of interest and reference annotaions may be different for different electrograms.

Also updated numpy and scipy versions. Required modifying the matlab reader switch from deprecated scipy functions. Set minimum scipy version to 1.8 to avoid needing to account for the breaking change.

In tests.test_case_routines the bipolar egms must be floats rather than ints, otherwise an error is thrown in get_sample_indices_within_woi when the samples outside the woi are set to np.NaN. They must be set to be NaN to allow for the calculation of the voltage.

+++ tests/test_case_routines.py
@@ -53,7 +53,7 @@ def mock_case(mocker):
     case.electric.annotations.window_of_interest = np.full((10, 2), fill_value=5)
     case.electric.annotations.window_of_interest[:, 1] += 10
 
-    case.electric.bipolar_egm.egm = np.repeat(np.arange(20)[np.newaxis, :], 10, axis=0)
+    case.electric.bipolar_egm.egm = np.repeat(np.arange(20, dtype=float)[np.newaxis, :], 10, axis=0)
     case.electric.internal_names = np.arange(10).astype(str)
 
     case.points = np.repeat(np.arange(20)[:, np.newaxis], repeats=3, axis=1)
